[FEATURE] CRUD Turnos and Security layer Done!!

+++ odontologia/src/main/java/com/proyectointegrador/odontologia/controller/PacienteController.java
@@ -5,6 +5,7 @@ import com.proyectointegrador.odontologia.dto.PacienteDto;
 import com.proyectointegrador.odontologia.dto.request.PacienteDtoReq;
 import com.proyectointegrador.odontologia.dto.response.PacienteDtoRes;
 import com.proyectointegrador.odontologia.entity.Paciente;
+import com.proyectointegrador.odontologia.exception.ResourceNotFoundException;
 import com.proyectointegrador.odontologia.service.IPacienteService;
 import com.proyectointegrador.odontologia.service.imp.PacienteServiceImp;
 import org.springframework.http.HttpStatus;
@@ -28,6 +29,11 @@ public class PacienteController {
         return new ResponseEntity<>(pacienteService.findAllPacientes(),HttpStatus.OK);
     }
 
+    @GetMapping("/findById/{id}")
+    public ResponseEntity<PacienteDto> findById(@PathVariable Long id){
+        return new ResponseEntity<>( pacienteService.findPacienteById2(id),HttpStatus.OK);
+    }
+
     @PostMapping("/crearPaciente")
     public ResponseEntity<PacienteDtoRes> crearPaciente(@RequestBody PacienteDtoReq pacienteDtoReq){
         return new ResponseEntity<>(pacienteService.createPaciente(pacienteDtoReq), HttpStatus.CREATED);
@@ -40,8 +46,14 @@ public class PacienteController {
     }
 
     @DeleteMapping("/deletePaciente/{id}")
-    public ResponseEntity<String> deletePaciente(@PathVariable Long id){
-        return  new ResponseEntity<>(pacienteService.deletePaciente(id),HttpStatus.OK);
+    public ResponseEntity<String> deletePaciente(@PathVariable Long id) throws ResourceNotFoundException {
+        ResponseEntity<String> response = null;
+        if (id != null){
+            response = new ResponseEntity<>(pacienteService.deletePaciente(id),HttpStatus.OK);
+        }else {
+            response = ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        return response;
     }
